deployment/requirements.txt
transformers==4.30.2
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asteroid==0.7.0
asteroid-filterbanks==0.4.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
braceexpand==0.1.7
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Distance==0.1.3
dnspython==2.6.1
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
filelock==3.13.1
fonttools==4.52.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
g2p-en==2.1.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.1
hydra-core==1.3.2
idna==3.7
inflect==7.2.1
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiwer==3.0.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
julius==0.2.7
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.5
lazy_loader==0.4
Levenshtein==0.25.1
librosa==0.10.2.post1
lightning-utilities==0.11.2
llvmlite==0.42.0
loguru==0.7.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mir_eval==0.7
mistune==3.0.2
mkl==2021.4.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nemo-toolkit==1.21.0
nest-asyncio==1.6.0
networkx==3.2.1
noisereduce==3.0.2
numba==0.59.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.1
opt-einsum==3.3.0
orjson==3.10.5
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pb-bss-eval==0.0.2
pesq==0.0.4
pickleshare==0.7.5
pillow==10.2.0
plac==1.4.3
platformdirs==4.2.2
pooch==1.8.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
pure-eval==0.2.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyarrow==16.1.0
pyasn1_modules==0.4.0
pybind11==2.12.0
pycparser==2.22
pydantic==1.10.15
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytorch-lightning==2.0.7
pytorch-ranger==0.1.1
pytz==2024.1
sentencepiece==0.1.97
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.1
wandb==0.17.0
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.6.0
webdataset==0.1.62
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wget==3.2
widgetsnbextension==4.0.11
wrapt==1.16.0
yarl==1.9.4
youtokentome==1.0.6
flask